brewlog/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brewlog"
version = "0.1.0"
description = "A local CLI brew tracker using the BrewSpec format"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jsonschema>=4.18",
]

[project.scripts]
brewlog = "brewlog.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
brewlog = ["brewspec.schema.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
